Create README.md
OneriSistemi/Öneri Sistemi is an  “Employee Suggestion System”. 
It is a solution that provides the monitoring and management of all suggestion processes over the system by acting as a bridge between the management of companies and employees. 
The suggestion system solution ensures that all suggestion processes are managed in the most effective way, thanks to artificial intelligence and machine learning technologies.